ref(preprod): Rename install_info.py to installable_builds.py
Consolidate public build distribution response models into a single
file, consistent with how size_analysis.py is organized.
